refactor(shell): extract normalizeHeadersInit to proxy-headers module
Agent-Id: agent-ec9c9e07-8486-4c98-8c4c-d579272b257d
Linked-Note-Id: 3d9e188f-1aa0-4bcc-80c3-ee3ec35b7ea2
